requirements.txt
# Requirements pour YouTube Shorts Generator
moviepy==1.0.3
yt-dlp>=2023.10.13
openai-whisper>=20231117
requests>=2.31.0
beautifulsoup4>=4.12.2
pydub>=0.25.1
numpy>=1.24.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
Pillow>=10.0.0